requirements.txt
python-dotenv
django
djangorestframework
pyjwt==1.7.1
python-keycloak